contrib/serving-simulator/requirements.txt
nvidia-ml-py
requests
paramiko
pyyaml
transformers
aiohttp
humanize